refactored drawer views to one fade
