+++ client/src/components/drawerviews/CardDrawerView.js
@@ -20,32 +20,23 @@ export default function ColumnDrawerView(props) {
   };
 
   return (
-    <Fade
-      right
-      when={context[0].open && context[0].type === 'card'}
-      collapse
-      duration={400}
+    <div
+      style={{
+        width: '100%',
+        display: 'flex',
+        justifyContent: 'space-between',
+        marginTop: 5,
+        padding: 3,
+      }}
     >
-      <div style={drawerStyle}>
-        <div
-          style={{
-            width: '100%',
-            display: 'flex',
-            justifyContent: 'space-between',
-            marginTop: 5,
-            padding: 3,
-          }}
-        >
-          <p>Card Edit</p>
-          <MdClose
-            size="1.5em"
-            style={{ cursor: 'pointer' }}
-            onClick={() => {
-              context[1]({ ...context[0], open: false });
-            }}
-          />
-        </div>
-      </div>
-    </Fade>
+      <p>Card Edit</p>
+      <MdClose
+        size="1.5em"
+        style={{ cursor: 'pointer' }}
+        onClick={() => {
+          context[1]({ ...context[0], open: false });
+        }}
+      />
+    </div>
   );
 }
