+++ client/src/components/drawerviews/SettingsDrawerView.js
@@ -4,7 +4,6 @@ import { GrHistory } from 'react-icons/gr';
 import Label from '../Label';
 import React, { useContext } from 'react';
 import { AutoContext } from '../../AutoContext';
-import Fade from 'react-reveal/Fade';
 
 export default function SettingsDrawerView(props) {
   const drawerStyle = {
@@ -57,79 +56,77 @@ export default function SettingsDrawerView(props) {
   const context = useContext(AutoContext);
 
   return (
-    <Fade right when={context[0].open} collapse duration={400}>
-      <div style={drawerStyle}>
-        <div
-          style={{
-            width: '100%',
-            display: 'flex',
-            justifyContent: 'space-between',
-            marginTop: 5,
-            padding: 3,
+    <>
+      <div
+        style={{
+          width: '100%',
+          display: 'flex',
+          justifyContent: 'space-between',
+          marginTop: 5,
+          padding: 3,
+        }}
+      >
+        <p>Settings</p>
+        <MdClose
+          size="1.5em"
+          style={{ cursor: 'pointer' }}
+          onClick={() => {
+            context[1]({ ...context[0], open: false });
           }}
-        >
-          <p>Settings</p>
-          <MdClose
-            size="1.5em"
-            style={{ cursor: 'pointer' }}
-            onClick={() => {
-              context[1]({ ...context[0], open: false });
-            }}
-          />
-        </div>
-        {/* title */}
-        <h5 style={{ alignSelf: 'center' }}>Project Autoban</h5>
-        {/* description */}
-        <div style={{ padding: 5, textAlign: 'center' }}>
-          <p style={{ color: 'gray' }}>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit,
-            sed do eiusmod tempor incididunt ut labore et dolore magna
-          </p>{' '}
-          <input
-            value="https://autobanprod.herokuapp.com/autobanproj"
-            style={{
-              border: '1px solid lightgray',
-              borderRadius: 10,
-            }}
-          />
-          <span style={{ fontSize: 10, color: 'lightgray' }}>
-            project url
-          </span>
-        </div>
-        <div style={linebreak}></div>
-        <div style={{ display: 'flex', justifyContent: 'center' }}>
-          <span>
-            <img
-              style={{ width: 30, height: 30, borderRadius: 5 }}
-              src="https://starwarsblog.starwars.com/wp-content/uploads/2019/08/d23-the-mandalorian-poster-tall-A-1088x816.jpg"
-            />{' '}
-          </span>{' '}
-          <button style={changeBGButton}>Change Background</button>
-        </div>
-
-        <div style={linebreak}></div>
-        <div style={rowFormat}>
-          <Label name="Backend" color="red" />
-          <Label name="Frontend" color="blue" />
-          <Label name="Testing" color="green" />
-          <Label name="Docs" color="purple" />
-
-          <p style={{ color: 'gray', fontSize: 12 }}>
-            Add Labels
-            <span>
-              <FaPlus />
-            </span>
-          </p>
-        </div>
-
-        <div style={linebreak}></div>
-        <div style={rowFormat}>
-          <span>
-            <GrHistory /> archived
-          </span>
-          <button style={closeButton}>Close Project</button>
-        </div>
+        />
       </div>
-    </Fade>
+      {/* title */}
+      <h5 style={{ alignSelf: 'center' }}>Project Autoban</h5>
+      {/* description */}
+      <div style={{ padding: 5, textAlign: 'center' }}>
+        <p style={{ color: 'gray' }}>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
+          do eiusmod tempor incididunt ut labore et dolore magna
+        </p>{' '}
+        <input
+          value="https://autobanprod.herokuapp.com/autobanproj"
+          style={{
+            border: '1px solid lightgray',
+            borderRadius: 10,
+          }}
+        />
+        <span style={{ fontSize: 10, color: 'lightgray' }}>
+          project url
+        </span>
+      </div>
+      <div style={linebreak}></div>
+      <div style={{ display: 'flex', justifyContent: 'center' }}>
+        <span>
+          <img
+            style={{ width: 30, height: 30, borderRadius: 5 }}
+            src="https://starwarsblog.starwars.com/wp-content/uploads/2019/08/d23-the-mandalorian-poster-tall-A-1088x816.jpg"
+          />{' '}
+        </span>{' '}
+        <button style={changeBGButton}>Change Background</button>
+      </div>
+
+      <div style={linebreak}></div>
+      <div style={rowFormat}>
+        <Label name="Backend" color="red" />
+        <Label name="Frontend" color="blue" />
+        <Label name="Testing" color="green" />
+        <Label name="Docs" color="purple" />
+
+        <p style={{ color: 'gray', fontSize: 12 }}>
+          Add Labels
+          <span>
+            <FaPlus />
+          </span>
+        </p>
+      </div>
+
+      <div style={linebreak}></div>
+      <div style={rowFormat}>
+        <span>
+          <GrHistory /> archived
+        </span>
+        <button style={closeButton}>Close Project</button>
+      </div>
+    </>
   );
 }
